+++ pos_orders_history_reprint/doc/changelog.rst
@@ -3,6 +3,16 @@
 Updates
 =======
 
+`1.0.2`
+-------
+
+- **Fix:** Error related to different localizations
+
+`1.0.1`
+-------
+
+- **IMP** Show order ticket after clicking the reprint button if the receipt of the order does not exist
+
 `1.0.0`
 -------